% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA_geology.R
\docType{data}
\name{CA_geology}
\alias{CA_geology}
\title{Geological map of California}
\format{
A SpatialPolygonsDataFrame
}
\usage{
data(CA_geology)
}
\description{
Attributes: `TYPE`, `DESC`, `TABLE`, `GENERAL`, `POR`, `QID`, `STATE`, `GEOLID`, `ID`
}
\references{
Jennings CW (1977) Geologic map of California. California Division of Mines and Geology geologic data. Map number 2, scale 1:750,000

Flint, L.E., Flint, A.L., Thorne, J.H., and Boynton, R., 2013, Fine-scale hydrologic modeling for regional landscape applications; the California Basin Characterization Model development and performance; Ecological Processes 2:25
}
\keyword{dataset}
