% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{feature_importance}
\alias{feature_importance}
\title{Compute feature importance}
\usage{
feature_importance(
  training_data,
  target_colname,
  filter_methods = c("FSelectorRcpp_information.gain"),
  .iters = 500,
  .first = 30,
  .split = 0.8,
  .stratify = TRUE,
  .seed = 1789
)
}
\arguments{
\item{training_data}{a `data.frame`}

\item{target_colname}{`character` the name of the column containing the target (output)}

\item{filter_methods}{`list` of `character` accepted by `mlr::generateFilterValuesData()`}

\item{.iters}{`numeric` number of iterations for the subsampling, default to 500}

\item{.first}{`numeric`, number of feature to display, default to 30}

\item{.split}{`numeric`, ratio of the subsampling splitting ratio, default to 0.8}

\item{.stratify}{`logical`, should the subsampling be stratified, default to `TRUE`}

\item{.seed}{`numeric`, fix seed for reproducible example}
}
\value{
a list with two elements a list of `data.frame` for each method and a list of `ggplot` objects
}
\description{
Compute feature importance
}
\keyword{function}
