% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_varnames.R
\name{fix_varnames}
\alias{fix_varnames}
\title{Pretty the name of variables}
\usage{
fix_varnames(df, var = "var")
}
\arguments{
\item{df}{a `data.frame`}

\item{var}{`character` the column of `df` where the variable id are}
}
\value{
a `data.frame`
}
\description{
Pretty the name of variables
}
\keyword{function}
