% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_feature_importance_plot.R
\name{make_feature_importance_plot}
\alias{make_feature_importance_plot}
\title{Makes a dot chart of feature importance}
\usage{
make_feature_importance_plot(feature_importance, first = 20, best = FALSE)
}
\arguments{
\item{feature_importance}{a `data.frame` with column `Overall` and `var`}

\item{first}{`numeric`, number of feature to display}

\item{best}{`logical`, default to `FALSE`, to highlight which features are selected in an optimal (best) set}
}
\value{
a `ggplot` object
}
\description{
Makes a dot chart of feature importance
}
\keyword{function}
