% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{plot_bootstrap_probability}
\alias{plot_bootstrap_probability}
\title{Visualize the probability result from bootstrapping analysis}
\usage{
plot_bootstrap_probability(l_boot, alpha = 0.95)
}
\arguments{
\item{l_boot}{a `list` generated with `bootstrap_freq()`}

\item{alpha}{`numeric`, a confidence level, default to `0.95`.}
}
\value{
a `ggplot` object
}
\description{
Visualize the probability result from bootstrapping analysis
}
\keyword{function}
