% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_data.R
\name{sanitize_data}
\alias{sanitize_data}
\title{Ensures that non-finite values are flagged as `NA`}
\usage{
sanitize_data(data_df)
}
\arguments{
\item{data_df}{a `data.frame`}
}
\value{
a `data.frame`
}
\description{
Ensures that non-finite values are flagged as `NA`
}
\keyword{function}
