% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significative_count_lollipop.R
\name{significative_count_lollipop}
\alias{significative_count_lollipop}
\title{Makes a lollipop chart from pair-wise comparison data}
\usage{
significative_count_lollipop(
  pairwise_results,
  pvalue = 0.05,
  normalize = FALSE,
  first = 25
)
}
\arguments{
\item{pairwise_results}{a `list` passed from `significative_diff_boxplot()`}

\item{pvalue}{`numeric` the threshold p-value}

\item{normalize}{`logical` if `TRUE`, normalize the count by the number of comparison}

\item{first}{`numeric` controls the number of variables plotted}
}
\value{
a named list with two elements `p` a `ggplot` plot and `stats` the count table of the significative value
}
\description{
Makes a lollipop chart from pair-wise comparison data
}
\keyword{function}
