% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SSCT_labelled_data}
\alias{SSCT_labelled_data}
\title{Statewide geospatial dataset}
\format{
A `data.frame`
}
\usage{
data(SSCT_labelled_data)
}
\description{
Statewide geospatial dataset
}
\keyword{dataset}
