% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{plot_pairwise_pvalue}
\alias{plot_pairwise_pvalue}
\title{Visualize the p-value result from the statistical pair-wise analysis}
\usage{
plot_pairwise_pvalue(stats, alpha = 0.95)
}
\arguments{
\item{alpha}{`numeric`, a confidence level, default to `0.95`.}

\item{l_boot}{a `list` generated with `bootstrap_freq()`}
}
\value{
a `ggplot` object
}
\description{
Visualize the p-value result from the statistical pair-wise analysis
}
\keyword{function}
